application/requirements.txt
streamlit
SPARQLWrapper
pandas
Pillow
